chore: restructure project with new app component and TypeScript configuration
