+++ package.json
@@ -16,7 +16,6 @@
 	},
 	"dependencies": {
 		"@tailwindcss/vite": "^4.1.18",
-		"@tc96/biome-config": "^2.2.0",
 		"react": "^19.2.3",
 		"react-dom": "^19.2.3",
 		"tailwindcss": "^4.1.18"
@@ -25,15 +24,18 @@
 		"@biomejs/biome": "2.3.10",
 		"@commitlint/cli": "20.2.0",
 		"@commitlint/config-conventional": "20.2.0",
+		"@tc96/biome-config": "^2.2.0",
+		"@types/node": "^25.0.3",
 		"@types/react": "^19.2.7",
 		"@types/react-dom": "^19.2.3",
-		"@vitejs/plugin-react": "^5.1.2",
+		"@vitejs/plugin-react-swc": "^4.2.2",
 		"autoprefixer": "10.4.23",
 		"globals": "^16.5.0",
 		"husky": "9.1.7",
 		"npm-check-updates": "^19.2.0",
 		"typescript": "~5.9.3",
-		"vite": "^7.3.0"
+		"vite": "^7.3.0",
+		"vite-tsconfig-paths": "^6.0.3"
 	},
 	"lint-staged": {
 		"src/**/*.{ts,tsx,js,jsx}": [
