+++ src/main.tsx
@@ -1,6 +1,6 @@
+import { App } from '@views/pages/app'
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import { App } from './components/App'
 
 // biome-ignore lint: noUnusedVariables
 createRoot(document.getElementById('root')!).render(
