+++ tsconfig.json
@@ -3,5 +3,6 @@
 	"references": [
 		{ "path": "./tsconfig.app.json" },
 		{ "path": "./tsconfig.node.json" }
-	]
+	],
+	 "exclude": ["node_modules"]
 }
