+++ vite.config.ts
@@ -1,7 +1,17 @@
-import tailwindcss from '@tailwindcss/vite'
-import react from '@vitejs/plugin-react'
 import { defineConfig } from 'vite'
 
+import tailwindcss from '@tailwindcss/vite'
+import react from '@vitejs/plugin-react-swc'
+import tsconfigPaths from 'vite-tsconfig-paths'
+
 export default defineConfig({
-	plugins: [react(), tailwindcss()],
+	plugins: [
+		react(),
+		tailwindcss(),
+		tsconfigPaths({ projects: ['tsconfig.app.json'] })
+	],
+	server: {
+		open: true,
+		port: 3000
+	},
 })
